Move Database to separate module
