+++ crates/core/src/lib.rs
@@ -1,26 +1,9 @@
+pub mod database;
 pub mod connection;
 pub mod errors;
 
+pub type Result<T> = std::result::Result<T, errors::Error>;
+
+pub use database::Database;
 pub use connection::Connection;
 pub use connection::ResultSet;
-
-use errors::Error;
-use std::ffi::c_int;
-
-type Result<T> = std::result::Result<T, Error>;
-
-pub struct Database {
-    pub url: String,
-}
-
-impl Database {
-    pub fn open<S: Into<String>>(url: S) -> Database {
-        Database { url: url.into() }
-    }
-
-    pub fn close(&self) {}
-
-    pub fn connect(&self) -> Result<Connection> {
-        Connection::connect(self)
-    }
-}
